Move _Thumbnail to Thumbnail
